create ReportErrorCriterium
