+++ confiture-web-app/src/components/report/ReportErrors.vue
@@ -1,17 +1,11 @@
 <script setup lang="ts">
-import { chunk } from "lodash-es";
-import { marked } from "marked";
 import { computed, ref } from "vue";
 
-import rgaa from "../../criteres.json";
 import { useReportStore } from "../../store";
 import { CriterionResultUserImpact, CriteriumResultStatus } from "../../types";
-import { formatStatus, formatUserImpact, getUploadUrl } from "../../utils";
-import CriteriumTestsAccordion from "../audit/CriteriumTestsAccordion.vue";
-import LazyAccordion from "../audit/LazyAccordion.vue";
-import MarkdownRenderer from "../ui/MarkdownRenderer.vue";
 import { getReportErrors, getReportTransverseErrors } from "./getReportErrors";
 import ReportCriteria from "./ReportCriteria.vue";
+import ReportErrorCriterium from "./ReportErrorCriterium.vue";
 
 const report = useReportStore();
 
@@ -74,21 +68,6 @@ const quickWinFilter = ref(false);
 function resetFilters() {
   userImpactFilters.value = defaultUserImpactFillters;
   quickWinFilter.value = false;
-}
-
-// Utility
-function getCriterium(topicNumber: number, criteriumNumber: number) {
-  // FIXME: "any everywhere" : The criteria properties of each topic do not have the same signature. See: https://github.com/microsoft/TypeScript/issues/33591#issuecomment-786443978
-  const criterium = (rgaa.topics as any)
-    .find((t: any) => t.number === topicNumber)
-    ?.criteria.find((c: any) => c.criterium.number === criteriumNumber)
-    .criterium;
-
-  return criterium;
-}
-
-function getCriteriumTitle(topicNumber: number, criteriumNumber: number) {
-  return marked.parseInline(getCriterium(topicNumber, criteriumNumber).title);
 }
 </script>
 
@@ -214,92 +193,13 @@ function getCriteriumTitle(topicNumber: number, criteriumNumber: number) {
           <p class="fr-tag fr-tag--sm fr-mb-3w">
             {{ topic.topic }}.&nbsp;{{ topic.name }}
           </p>
-          <template v-for="(error, j) in topic.errors" :key="j">
-            <p
-              :class="[
-                'fr-text--lg fr-text--bold criterium-title',
-                { 'fr-mt-9v': j !== 0 }
-              ]"
-            >
-              {{ error.topic }}.{{ error.criterium }}&nbsp;
-              <span v-html="getCriteriumTitle(error.topic, error.criterium)" />
-            </p>
 
-            <ul class="fr-badges-group fr-mb-2w">
-              <li>
-                <p
-                  class="fr-badge fr-badge--sm fr-badge--error fr-badge--no-icon"
-                >
-                  {{ formatStatus(error.status) }}
-                </p>
-              </li>
-              <li v-if="error.userImpact">
-                <p
-                  class="fr-badge fr-badge--sm"
-                  :class="{
-                    'fr-badge--yellow-moutarde':
-                      error.userImpact === CriterionResultUserImpact.MAJOR,
-                    'fr-badge--error fr-badge--no-icon':
-                      error.userImpact === CriterionResultUserImpact.BLOCKING
-                  }"
-                >
-                  Impact {{ formatUserImpact(error.userImpact) }}
-                </p>
-              </li>
-              <li v-if="error.quickWin">
-                <p class="fr-badge fr-badge--sm">Facile à corriger</p>
-              </li>
-            </ul>
-
-            <!-- Error -->
-            <LazyAccordion
-              v-if="error.notCompliantComment || error.exampleImages.length > 0"
-              title="Erreur et recommandation"
-              data-accordion
-            >
-              <MarkdownRenderer
-                v-if="error.notCompliantComment"
-                class="fr-mb-3w"
-                :markdown="error.notCompliantComment"
-              />
-              <p
-                v-if="chunk(error.exampleImages, 2).length"
-                class="fr-text--xs fr-mb-1w error-accordion-subtitle"
-              >
-                Exemple(s) d’erreur(s)
-              </p>
-              <div class="fr-container--fluid">
-                <div
-                  v-for="(line, k) in chunk(error.exampleImages, 2)"
-                  :key="k"
-                  class="fr-grid-row fr-grid-row--gutters"
-                >
-                  <a
-                    v-for="example in line"
-                    :key="example.key"
-                    class="fr-col-md-6 fr-col-12 image-link"
-                    :href="getUploadUrl(example.key)"
-                    target="_blank"
-                  >
-                    <span class="sr-only">
-                      Ouvrir l’image dans une nouvelle fenêtre
-                    </span>
-                    <img
-                      style="width: 100%"
-                      :src="getUploadUrl(example.key)"
-                      alt=""
-                    />
-                  </a>
-                </div>
-              </div>
-            </LazyAccordion>
-
-            <!-- Tests -->
-            <CriteriumTestsAccordion
-              :topic-number="error.topic"
-              :criterium="getCriterium(error.topic, error.criterium)"
-            />
-          </template>
+          <ReportErrorCriterium
+            v-for="(error, j) in topic.errors"
+            :key="j"
+            :error="error"
+            :class="{ 'fr-mt-9v': j !== 0 }"
+          />
         </div>
       </section>
     </template>
@@ -338,92 +238,13 @@ function getCriteriumTitle(topicNumber: number, criteriumNumber: number) {
           <p class="fr-tag fr-tag--sm fr-mb-3w">
             {{ topic.topic }}.&nbsp;{{ topic.name }}
           </p>
-          <template v-for="(error, j) in topic.errors" :key="j">
-            <p
-              :class="[
-                'fr-text--lg fr-text--bold criterium-title',
-                { 'fr-mt-9v': j !== 0 }
-              ]"
-            >
-              {{ error.topic }}.{{ error.criterium }}&nbsp;
-              <span v-html="getCriteriumTitle(error.topic, error.criterium)" />
-            </p>
 
-            <ul class="fr-badges-group fr-mb-2w">
-              <li>
-                <p
-                  class="fr-badge fr-badge--sm fr-badge--error fr-badge--no-icon"
-                >
-                  {{ formatStatus(error.status) }}
-                </p>
-              </li>
-              <li v-if="error.userImpact">
-                <p
-                  class="fr-badge fr-badge--sm"
-                  :class="{
-                    'fr-badge--yellow-moutarde':
-                      error.userImpact === CriterionResultUserImpact.MAJOR,
-                    'fr-badge--error fr-badge--no-icon':
-                      error.userImpact === CriterionResultUserImpact.BLOCKING
-                  }"
-                >
-                  Impact {{ formatUserImpact(error.userImpact) }}
-                </p>
-              </li>
-              <li v-if="error.quickWin">
-                <p class="fr-badge fr-badge--sm">Facile à corriger</p>
-              </li>
-            </ul>
-
-            <!-- Error -->
-            <LazyAccordion
-              v-if="error.notCompliantComment || error.exampleImages.length > 0"
-              title="Erreur et recommandation"
-              data-accordion
-            >
-              <MarkdownRenderer
-                v-if="error.notCompliantComment"
-                class="fr-mb-3w"
-                :markdown="error.notCompliantComment"
-              />
-              <p
-                v-if="chunk(error.exampleImages, 2).length"
-                class="fr-text--xs fr-mb-1w error-accordion-subtitle"
-              >
-                Exemple(s) d’erreur(s)
-              </p>
-              <div class="fr-container--fluid">
-                <div
-                  v-for="(line, k) in chunk(error.exampleImages, 2)"
-                  :key="k"
-                  class="fr-grid-row fr-grid-row--gutters"
-                >
-                  <a
-                    v-for="example in line"
-                    :key="example.key"
-                    class="fr-col-md-6 fr-col-12 image-link"
-                    :href="getUploadUrl(example.key)"
-                    target="_blank"
-                  >
-                    <span class="sr-only">
-                      Ouvrir l’image dans une nouvelle fenêtre
-                    </span>
-                    <img
-                      style="width: 100%"
-                      :src="getUploadUrl(example.key)"
-                      alt=""
-                    />
-                  </a>
-                </div>
-              </div>
-            </LazyAccordion>
-
-            <!-- Tests -->
-            <CriteriumTestsAccordion
-              :topic-number="error.topic"
-              :criterium="getCriterium(error.topic, error.criterium)"
-            />
-          </template>
+          <ReportErrorCriterium
+            v-for="(error, j) in topic.errors"
+            :key="j"
+            :error="error"
+            :class="{ 'fr-mt-9v': j !== 0 }"
+          />
         </div>
       </section>
     </template>
